cleanup all for resources (#339)
* cleanup all for resources


* fix code smell and add tests


* update readme


* update readme

+++ controllers/restore_post.go
@@ -339,6 +339,10 @@ func deleteDynamicResourcesForBackup(
 	}
 	labelSelector := fmt.Sprintf("%s, %s notin (%s), %s",
 		BackupNameVeleroLabel, BackupNameVeleroLabel, backupName, genericLabel)
+	if restoreOptions.cleanupType == v1beta1.CleanupTypeAll {
+		// get all resources, including user created
+		labelSelector = genericLabel
+	}
 	if otherLabels != "" {
 		labelSelector = fmt.Sprintf("%s, %s", labelSelector, otherLabels)
 	}
@@ -370,21 +374,46 @@ func deleteDynamicResourcesForBackup(
 				groupKind, err.Error()))
 			continue
 		}
-		if dr := restoreOptions.dynamicArgs.dyn.Resource(mapping.Resource); dr != nil {
+		invokeDynamicDelete(ctx, restoreOptions, labelSelector,
+			veleroBackup, mapping)
+	}
 
-			listOptions := v1.ListOptions{LabelSelector: labelSelector}
-			if dynamiclist, err := dr.List(ctx, listOptions); err == nil {
-				// get all items and delete them
-				for i := range dynamiclist.Items {
-					deleteDynamicResource(
-						ctx,
-						mapping,
-						dr,
-						dynamiclist.Items[i],
-						restoreOptions.deleteOptions,
-						veleroBackup.Spec.ExcludedNamespaces,
-					)
+}
+
+func invokeDynamicDelete(
+	ctx context.Context,
+	restoreOptions RestoreOptions,
+	labelSelector string,
+	veleroBackup *veleroapi.Backup,
+	mapping *meta.RESTMapping,
+) {
+
+	backupName := veleroBackup.Name
+	if dr := restoreOptions.dynamicArgs.dyn.Resource(mapping.Resource); dr != nil {
+
+		var listOptions = v1.ListOptions{}
+		if labelSelector != "" {
+			listOptions = v1.ListOptions{LabelSelector: labelSelector}
+		}
+		if dynamiclist, err := dr.List(ctx, listOptions); err == nil {
+			// get all items and delete them
+			for i := range dynamiclist.Items {
+
+				item := dynamiclist.Items[i]
+				if restoreOptions.cleanupType == v1beta1.CleanupTypeAll &&
+					item.GetLabels()[BackupNameVeleroLabel] == backupName {
+					// exclude here resources with the same backup as the last restore
+					continue
 				}
+
+				deleteDynamicResource(
+					ctx,
+					mapping,
+					dr,
+					item,
+					restoreOptions.deleteOptions,
+					veleroBackup.Spec.ExcludedNamespaces,
+				)
 			}
 		}
 	}
